+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * feat(Editor): handle boolean in `image` and `mention` props
 * chore(deps): update tiptap to ^3.15.0
 * docs(table): improve examples to use meta
+* feat(Editor): add `placeholder.mode` prop
 
 ## [2.2.1](https://github.com/bitrix24/b24ui/compare/v2.1.17...v2.2.1) (2025-12-18)